Reapply "Break circular dependencies from file_resources.ts (#20485)" (#20498)
This reverts commit 68cb09d6b0dc25b1f60f772b5831d70191a42c8d.
