+++ front/lib/actions/mcp_internal_actions/utils/file_utils.ts
@@ -11,11 +11,11 @@ import {
 } from "@app/lib/api/assistant/conversation/attachments";
 import type { Authenticator } from "@app/lib/auth";
 import { FileResource } from "@app/lib/resources/file_resource";
+import { streamToBuffer } from "@app/lib/utils/streams";
 import {
   isAgentMessageType,
   isContentFragmentType,
   isInteractiveContentFileContentType,
-  normalizeError,
 } from "@app/types";
 
 export function sanitizeFilename(filename: string): string {
@@ -123,24 +123,4 @@ export async function getFileFromConversationAttachment(
     contentType: attachment.contentType || "application/octet-stream",
     fileResource,
   });
-}
-
-export async function streamToBuffer(
-  readStream: NodeJS.ReadableStream
-): Promise<Result<Buffer, string>> {
-  try {
-    const chunks: Buffer[] = [];
-    for await (const chunk of readStream) {
-      if (Buffer.isBuffer(chunk)) {
-        chunks.push(chunk);
-      } else {
-        chunks.push(Buffer.from(chunk));
-      }
-    }
-    return new Ok(Buffer.concat(chunks));
-  } catch (error) {
-    return new Err(
-      `Failed to read file stream: ${normalizeError(error).message}`
-    );
-  }
 }
