+++ front/lib/api/actions/servers/image_generation/tools/index.ts
@@ -22,6 +22,7 @@ import {
 import { IMAGE_GENERATION_TOOLS_METADATA } from "@app/lib/api/actions/servers/image_generation/metadata";
 import type { Authenticator } from "@app/lib/auth";
 import { getStatsDClient } from "@app/lib/utils/statsd";
+import { streamToBuffer } from "@app/lib/utils/streams";
 import logger from "@app/logger/logger";
 import { Err, normalizeError } from "@app/types";
 import { GEMINI_3_PRO_IMAGE_MODEL_ID } from "@app/types/assistant/models/google_ai_studio";
