+++ front/lib/api/assistant/conversation/content_fragment.ts
@@ -11,8 +11,8 @@ import assert from "assert";
 import { isContentFragmentDataSourceNode } from "@app/lib/api/assistant/conversation/attachments";
 import config from "@app/lib/api/config";
 import { getContentNodeFromCoreNode } from "@app/lib/api/content_nodes";
-import type { ProcessAndStoreFileError } from "@app/lib/api/files/upload";
-import { processAndStoreFile } from "@app/lib/api/files/upload";
+import type { ProcessAndStoreFileError } from "@app/lib/api/files/processing";
+import { processAndStoreFile } from "@app/lib/api/files/processing";
 import type { Authenticator } from "@app/lib/auth";
 import { DataSourceViewResource } from "@app/lib/resources/data_source_view_resource";
 import { FileResource } from "@app/lib/resources/file_resource";
